Refactor auth logic into AuthService
